requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gspread==6.0.0
oauth2client==4.1.3
pydantic==2.6.0
pydantic-settings==2.1.0
razorpay==1.3.0
twilio==8.1.0
openai==1.10.0
google-generativeai==0.3.2
python-dotenv==1.0.1
httpx==0.26.0
